postprocess/requirements.txt
absl-py==2.3.1
beautifulsoup4==4.13.5
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.1.3
filelock==3.16.1
flake8==7.1.2
Flask==3.0.3
fonttools==4.57.0
fsspec==2025.3.0
future==1.0.0
gdown==5.2.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
isort==4.3.21
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lap==0.5.12
loguru==0.7.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
mpmath==1.3.0
mysql-connector-python==9.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
sympy==1.13.3
tabulate==0.9.0
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
win32_setctime==1.2.0
yacs==0.1.8
yapf==0.43.0
zipp==3.20.2
